added global window stuffs so we don't need to ts-ignore some part of the script

+++ src/data/sql.mongodb.tsx
@@ -6,8 +6,7 @@ const MONGO_ADAPTER_PREFIX = 'db';
 export function getSelectAllColumns(input: SqlAction.TableInput): SqlAction.Output | undefined {
   const label = `Select All Columns`;
 
-  switch (input.dialect) {
-    case 'mongodb':
+  if (input.dialect === 'mongodb') {
       return {
         label,
         formatter: 'sql',
@@ -28,20 +27,18 @@ export function getSelectSpecificColumns(
   const columnString = `\n` + input.columns.map((col) => `  ${col.name}`).join(',\n');
   const whereColumnString = input.columns.map((col) => `${col.name} = ''`).join('\n -- AND ');
 
-  switch (input.dialect) {
-    case 'mongodb':
-      const insertValueObject = {};
+  if (input.dialect === 'mongodb') {
+      const columns : any= {};
       for (const column of input.columns || []) {
         if (!column.name.includes('.')) {
-          //@ts-ignore
-          insertValueObject[column.name] = column.type === 'string' ? '' : 123;
+          columns[column.name] = column.type === 'string' ? '' : 123;
         }
       }
       return {
         label,
         formatter: 'sql',
         query: `${MONGO_ADAPTER_PREFIX}.collection('${input.tableId}').find(
-          ${JSON.stringify(insertValueObject)}
+          ${JSON.stringify(columns)}
         ).limit(${input.querySize}).toArray();`,
       };
   }
@@ -57,20 +54,18 @@ export function getInsertCommand(input: SqlAction.TableInput): SqlAction.Output 
   const columnString = input.columns.map((col) => col.name).join(',\n');
   const insertValueString = input.columns.map((col) => `'_${col.name}_'`).join(',\n');
 
-  switch (input.dialect) {
-    case 'mongodb':
-      const insertValueObject = {};
+  if (input.dialect === 'mongodb') {
+      const columns : any= {};
       for (const column of input.columns) {
         if (column.name !== '_id' && !column.name.includes('.')) {
-          //@ts-ignore
-          insertValueObject[column.name] = column.type === 'string' ? '' : 123;
+          columns[column.name] = column.type === 'string' ? '' : 123;
         }
       }
       return {
         label,
         formatter: 'js',
         query: `${MONGO_ADAPTER_PREFIX}.collection('${input.tableId}').insertMany([
-          ${JSON.stringify(insertValueObject)}
+          ${JSON.stringify(columns)}
         ]);`,
       };
   }
@@ -86,21 +81,19 @@ export function getUpdateCommand(input: SqlAction.TableInput): SqlAction.Output 
   const columnString = input.columns.map((col) => `-- ${col.name} = ''`).join(',\n');
   const whereColumnString = input.columns.map((col) => `-- ${col.name} = ''`).join(' AND \n');
 
-  switch (input.dialect) {
-    case 'mongodb':
-      const insertValueObject = {};
+  if (input.dialect === 'mongodb') {
+      const columns : any= {};
       for (const column of input.columns) {
         if (column.name !== '_id' && !column.name.includes('.')) {
-          //@ts-ignore
-          insertValueObject[column.name] = column.type === 'string' ? '' : 123;
+          columns[column.name] = column.type === 'string' ? '' : 123;
         }
       }
       return {
         label,
         formatter: 'js',
         query: `${MONGO_ADAPTER_PREFIX}.collection('${input.tableId}').update(
-          ${JSON.stringify(insertValueObject)},
-          {\$set: ${JSON.stringify(insertValueObject, null, 2)}}
+          ${JSON.stringify(columns)},
+          {\$set: ${JSON.stringify(columns, null, 2)}}
         );`,
       };
   }
@@ -115,20 +108,18 @@ export function getDeleteCommand(input: SqlAction.TableInput): SqlAction.Output 
 
   const whereColumnString = input.columns.map((col) => `-- ${col.name} = ''`).join(' AND \n');
 
-  switch (input.dialect) {
-    case 'mongodb':
-      const insertValueObject = {};
+  if (input.dialect === 'mongodb') {
+      const columns : any= {};
       for (const column of input.columns) {
         if (column.name !== '_id' && !column.name.includes('.')) {
-          //@ts-ignore
-          insertValueObject[column.name] = column.type === 'string' ? '' : 123;
+          columns[column.name] = column.type === 'string' ? '' : 123;
         }
       }
       return {
         label,
         formatter: 'js',
         query: `${MONGO_ADAPTER_PREFIX}.collection('${input.tableId}').deleteMany(
-          ${JSON.stringify(insertValueObject)}
+          ${JSON.stringify(columns)}
         );`,
       };
   }
@@ -144,8 +135,7 @@ export function getCreateTable(input: SqlAction.TableInput): SqlAction.Output | 
   let columnString: string = '';
 
   // TODO: figure out how to use the defaultval
-  switch (input.dialect) {
-    case 'mongodb':
+  if (input.dialect === 'mongodb') {
       return {
         label,
         formatter: 'js',
@@ -157,8 +147,7 @@ export function getCreateTable(input: SqlAction.TableInput): SqlAction.Output | 
 export function getDropTable(input: SqlAction.TableInput): SqlAction.Output | undefined {
   const label = `Drop Table`;
 
-  switch (input.dialect) {
-    case 'mongodb':
+  if (input.dialect === 'mongodb') {
       return {
         label,
         formatter: 'js',
